Support full config file
Closes #3 .

+++ waitcomment.sh
@@ -20,30 +20,42 @@ if [ -z $mytmpdir ]; then
 fi
 # Grabs the last ten lines of the sudo logfile (my location) to tmpfile.
 # Would be better to search from end of file but I don't know how to do that.
-tail -n10 $sudolog > $mytmpdir/comment3.tmp
-tac $mytmpdir/comment3.tmp | awk '!flag; /TTY/{flag = 1};' | tac > $mytmpdir/comment2.tmp
-rm $mytmpdir/comment3.tmp
-curr_command="$(grep -o COMMAND=.* $mytmpdir/comment2.tmp)"
-curr_shell="$(head -n1 $mytmpdir/comment2.tmp | cut -c 31-35)"
-# Creates a named pipe to communicate with 'child' addcomment process. Completely unnecessary
-# but I've never used named pipes before and thought I might as well learn how.
-mkfifo $tmpdir/$curr_shell
-# List of commands to be commented. Can be customized. Will soon be grabbed from config file rather than this mess. 
-if grep -qE '=/usr/bin/visudo|=/usr/bin/vim|=/usr/bin/rm|=/usr/bin/rmdir|=/usr/bin/mkdir|=/usr/bin/trash-put|=/usr/bin/cp|=/usr/bin/touch|=/usr/bin/mv|=/usr/bin/chmod|=/usr/bin/chown|=/usr/bin/chgrp|=/usr/bin/systemctl edit|=/usr/bin/pacman -S |=/usr/bin/pacman -R|>' <<< $curr_command; then
+tail -n10 $sudolog > $mytmpdir/comment_pre.tmp
+tac $mytmpdir/comment_pre.tmp | awk '!flag; /TTY/{flag = 1};' | tac > $mytmpdir/comment.tmp
+rm $mytmpdir/comment_pre.tmp
+curr_command="$(grep -o COMMAND=.* $mytmpdir/comment.tmp)"
+curr_shell="$(head -n1 $mytmpdir/comment.tmp | cut -c 31-35)"
+# Creates a named pipe to communicate with 'child' addcomment process.
+if [ -f $tmpdir/$curr_shell ]; then
+	cat $tmpdir/$curr_shell
+	echo "EXIT" > $tmpdir/$curr_shell
+else
+	mkfifo $tmpdir/$curr_shell
+fi
+# Parses config file options to pass to grep later.
+p_track=`printf '%s|' "|${track[@]}" | head -c -2 | sed 's/|/|COMMAND=\/usr\/bin\//g' | tail -c +2`
+p_exclude=`printf '%s|' "|${exclude[@]}" | head -c -2 | sed 's/|/|COMMAND=\/usr\/bin\//g' | tail -c +2`
+p_shell=`printf '%s|' "${shell[@]}" | head -c -1`
+# Tries to match run command to list in config file. ">" is included by default.
+if grep -qE "$p_track|>" <<< $curr_command; then
 	# Keeps sudo-comment from triggering itself in certain situations.
- 	# The pacman exclusions are to prevent makepkg from triggering this, because it causes problems (makepkg doesn't maintain a foreground process being that it's a 
-  	# shell script calling pacman several times). Use -Run for manual removals instead if you normally remove packages with -Rnu.
-	if grep -qE -v 'comment.tmp|comment2.tmp|comment3.tmp|changed.tmp|pacman -Rnu|pacman -S --asdeps' <<< $curr_command; then
+	# By default this includes all edits to $tmpdir and the specific pacman commands run by makepkg.
+	if grep -qE -v "$p_exclude|$tmpdir" <<< $curr_command; then
 		sleep 0.2
-		# Waits until no processes running in foreground on given shell except bash or sudo. Causes some problems with files named bash (ex. /etc/bash.bashrc). Should use something
-		# other than grep but I'll figure it out later.
-		until [ `ps ax | grep -v "bash" | grep -v "sudo -s" | grep -v "sudo -i" | awk '$3 ~ /\+/' | awk -v shell=$curr_shell '$2 ~ shell' | wc -l` == 0 ]
+		# Waits until no processes running in foreground on given shell except the shell itself.
+  		until [ `ps ax | awk -v c_shell=$curr_shell '$2 ~ c_shell' | awk '$3 ~ /\+/' | awk '{print $5}' | grep -vE -e "\$p_shell" |  wc -l` == 0 ]
 		do 
 			sleep 0.1
 		done
-  		cat $mytmpdir/comment2.tmp > $mytmpdir/comment.tmp
+  		# addcomment runs (normally) as user so these files need be to world-writable.
 		chmod 666 $mytmpdir/comment.tmp
 		chmod 666 $tmpdir/$curr_shell
+		ttyecho -n /dev/$curr_shell 'addcomment'
+		echo $mytmpdir > $tmpdir/$curr_shell
+		output=`cat $tmpdir/$curr_shell`
+		if [ $output = "OK" ]; then
+			cat $mytmpdir/comment.tmp >> $commentlog
+		fi
 		# Pushes "addcomment" with a newline to the shell.
 		ttyecho -n /dev/$curr_shell 'addcomment'
 		# Waits until addcomment reads pipe, then waits for return traffic.
@@ -53,9 +65,11 @@ if grep -qE '=/usr/bin/visudo|=/usr/bin/vim|=/usr/bin/rm|=/usr/bin/rmdir|=/usr/b
 		if [ $output = "OK" ]; then
 			cat $mytmpdir/comment.tmp >> $commentlog
 		fi
-		rm $mytmpdir/comment.tmp
 	fi
-fi	
-rm $mytmpdir/comment2.tmp
+fi
+# Cleans up. If the process was made to exit by a duplicate process, don't delete the pipe.
+rm $mytmpdir/comment.tmp
 rmdir $mytmpdir
-rm $tmpdir/$curr_shell
+if [ $output != "EXIT" ]; then
+	rm $tmpdir/$curr_shell
+fi
